refactor: extract explorer helpers + layout into submodules
explorer.rs was 1428 LOC mixing:
- HTML-escape + truncate + srx + addr-with-label + fmt-ts + fmt-day
  (pure formatters with no shared state)
- CSS + chart-section script + page wrapper + nav-tabs (chrome layout)
- Per-page handlers (home / blocks / transactions / block / address /
  tx / validators / tokens / richlist / validator / token / mempool +
  stats_daily endpoint)

Move the formatters into explorer/helpers.rs and the chrome into
explorer/layout.rs. explorer.rs drops to 1197 LOC, focused on the
handlers themselves. No behaviour change; cargo clippy + cargo test
both clean.

Phase 2 of the workspace audit (see internal AUDIT_REPORT). Per-handler
splits stay deferred — they'd add new module surface for marginal
readability gain on this round.
